Rework and improve downloading and output handling
Replace the Downloader class by a Storage class that handles both
downloading and storage on disk. All components have a storage which
allows to retrieve data from disk and download it if needed.

packagemanifest, clientmanifest and solutionmanifest are now stored on
disk. This will allow to analyze dependencies without having to
re-download anything.