Rework handling of project version files
Remove BinPackageFile. BinPackageFile now simply stores information from
packagemanifest; it also pre-compute compressed and extract_path as they
are used whenever a package file is used.

The parsed BinPackageFile instances are now cached on the
ProjectVersion.

Optimize BinPackageFile, as it is heavily used to retrieve patch
versions.

These changes reduced processing needed when searching for a single file
in a project version. which is used when retrieving patch versions.