Update 'patcher' to use the new release information URLs
The storage layout for release information changed and is not compatible
with the previous one. Bundles and files layout didn't changed.

Since there is no common versionning (and no available versionning for
client), manifest URLs of both game and client are compared to detect
when there is a new release. The version "number" is not its retrieval
timestamp (which may look weird when printed).

This is a working but transitional solution to have the latest patch
updated on cdragon. The design would benefit from a larger refactoring
to separate the retrieval of release information from the actual storage
part. Also, it would be nice to separate game and client components
completely, even if they are still merged when exported to cdragon.