Fix TFT character 'cost'
Cost can now usually be retrieved from 'tier'.
It is not always set and was not in previous versions, so keep the
fallback to the previous method.

The rarity value 7 can now be found here for some characters without a
tier. They usually are "special" characters, still considered as
champions (which is probably wrong).
This could probably be improved.